dat1-deploy-custom-models/llama-chat/requirements.txt
fastapi
requests
sse-starlette
sseclient-py
